% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_area_id.R
\name{add_area_id}
\alias{add_area_id}
\title{}
\usage{
add_area_id(df, df_areas_wide, par, add_keep_cols = NULL)
}
\arguments{
\item{df}{Dataframe with \code{area_id} column.}

\item{df_areas_wide}{\code{sf} \code{dataframe} with shapefiles and area
hierarchy.}

\item{par}{list with two entries:
\itemize{
\item{\code{area_lev}}{Current area level of \code{df}.}
\item{\code{area_lev_select}}{Desired area level for \code{df}.}
}}

\item{add_keep_cols}{Additional columns to keep when summarising,
Default: NULL}
}
\value{
\code{df} with \code{area_id} changed to \code{area_lev_select}.
}
\description{
Function to change \code{area_id} from one hierarchy level to
another.
}
