% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_mc_name.R
\name{append_mc_name}
\alias{append_mc_name}
\title{Append "mc" to Appropriate Column names}
\usage{
append_mc_name(.data)
}
\arguments{
\item{.data}{Dataframe/tibble whose columns include "MC" calculations.}
}
\value{
\code{.data}, with column names appended appropriately.
}
\description{
Ensure names for MC columns in fit have the suffix "_mc"
}
