% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{brackify}
\alias{brackify}
\title{Convert vector to string}
\usage{
brackify(x, quote = FALSE)
}
\description{
convert a vector like c(1, 4, 3, 2) into a string like
\verb{[1, 4, 3, 2]} (common aggregation method for error messages). See also
\code{data.table:::brackify()}.
}
\keyword{internal}
