% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{change_agegroup_convention}
\alias{change_agegroup_convention}
\title{Change age group convention to match aggregation results}
\usage{
change_agegroup_convention(.data)
}
\arguments{
\item{.data}{\code{data.frame} with \code{age_group} column whose convention does not
match aggregations from \code{threemc}.}
}
\description{
Change age group convention from "Y000_004" to "0-4", for
example.
}
\keyword{internal}
