% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_sample_tmb.R
\name{circ_sample_tmb}
\alias{circ_sample_tmb}
\title{Sample TMB fit for Circumcision Model}
\usage{
circ_sample_tmb(obj, opt, nsample = 1000, ...)
}
\arguments{
\item{obj}{TMB object/AD model outputted by \link[TMB]{MakeADFun}.}

\item{opt}{Optimised TMB model, outputted by optimisation function such
as \link[stats]{nlminb} or \link[stats]{optim}.}

\item{...}{Further arguments passed to or from other methods.}

\item{n_samples}{Number of samples to be generated. Additional parameters
for \link[naomi]{sample_tmb} can also be supplied, Default: 1000}
}
\value{
Object of class "naomi_fit", containing the original TMB object
("obj"), the standard deviation report for optimised AD model ("sdreport")
and \code{n_samples} samples for the (cumulative) incidence and hazard rate of
circumcision for the region(s) in question.
}
\description{
Sample from TMB object, using \link[naomi]{sample_tmb}. Saves
changing object to "Naomi" format. Also produces and returns standard
deviation report outputted by \link[TMB]{sdreport}.
}
\seealso{
\code{\link[TMB]{sdreport}}
\code{\link[naomi]{sample_tmb}}
}
