% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cmc_date_to_year_num}
\alias{cmc_date_to_year_num}
\title{Convert integer CMC date to integer Julian calendar year}
\usage{
cmc_date_to_year_num(cmc_date)
}
\arguments{
\item{cmc_date}{integer specifying date as CMC (Century-month code)}
}
\value{
integer Julian calendar year equivalent to CMC date.
}
\description{
Convert integer CMC date to integer Julian calendar year
}
\keyword{internal}
