% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{combine_areas}
\alias{combine_areas}
\title{Collect results for lower area hierarchies}
\usage{
combine_areas(.data, areas_wide, area_lev, join, add_keep_cols = NULL, ...)
}
\arguments{
\item{.data}{Results for highest area hierarchy, to be combined to
give results for lower/less granular area hierarchies.}

\item{areas_wide}{\code{data.frame} with shapefiles and area hierarchy.}

\item{area_lev}{Desired admin boundary level.}

\item{join}{Indicator to decide whether to join data for different
area hierarchies, or return them in list form.}

\item{add_keep_cols}{Additional columns to keep when summarising,
Default: NULL}

\item{...}{Further arguments passed to \link[threemc]{add_area_id}.}
}
\value{
\code{data.frame} or list (depending on the value of \code{join})
with results for all area levels less than or equal to \code{area_lev}.
}
\description{
Function to collect results for lower area hierarchies by
joining higher area hierarchies.
}
\keyword{internal}
