% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tmb.R
\name{compile_tmb}
\alias{compile_tmb}
\title{Compile TMB model without C++ Warnings}
\usage{
compile_tmb(file, logfile = "/tmp/tmb_logfile.log", ...)
}
\arguments{
\item{file}{C++ file.}

\item{logfile}{File to redirect output from C++ compiler. To print output
to R console, specify \code{logfile = NULL}, Default: '/tmp/tmb_logfile.log'}

\item{...}{Additional parameters for \code{\link[TMB]{compile}}.}
}
\description{
Wrapper around \code{\link[TMB]{compile}} which optionally
redirects C++ compiler output to a logfile, avoiding your R console, since
it can be quite verbose. Note that \code{logfile} argument default is not
valid for Windows.
}
\seealso{
\code{\link[TMB]{compile}}
}
