% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantiles.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{Calculate Quantiles for Rates and Cumulative Hazard}
\usage{
compute_quantiles(
  out,
  fit,
  area_lev = NULL,
  probs = c(0.025, 0.5, 0.975),
  names = FALSE,
  ...
)
}
\arguments{
\item{out}{Shell dataset with a row for every unique record in
circumcision survey data for a given area. Also includes empirical estimates
for circumcision estimates for each unique record.}

\item{fit}{Object containing \code{samples} list for the (cumulative) incidence
and hazard rate of circumcision for the region(s) in question.}

\item{probs}{Specific quantiles to be calculated,
Default: c(0.025, 0.5, 0.975)}

\item{names}{Parameter with \link[stats]{quantile}: logical; if true, the
result has a names attribute. Set to FALSE for speedup with many probs,
Default: FALSE}

\item{...}{Further arguments passed to \link[stats]{quantile}.}
}
\value{
Input \code{out} \code{data.frame}, including columns with quantiles for
hazard rates etc for different circumcision types, and for overall
circumcision.
}
\description{
Calculate quantiles for samples of rates and cumulative hazard
outputted from \link[threemc]{circ_sample_tmb}, and add them as columns to
the shell \code{data.frame} \code{out} with estimated empirical circumcision rates.
}
\seealso{
\code{\link[threemc]{circ_sample_tmb}}
\code{\link[stats]{quantile}}
@importFrom dplyr \%>\%
@importFrom rlang .data
}
