% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{create_data_indicators}
\alias{create_data_indicators}
\title{Create data indicators for TMB model}
\usage{
create_data_indicators(
  out,
  rw_order = NULL,
  paed_age_cutoff = NULL,
  inc_time_tmc = NULL
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{rw_order}{Order of the random walk used for temporal precision matrix.
Setting to NULL assumes you wish to specify an AR 1 temporal prior.
Default: NULL}

\item{paed_age_cutoff}{Age at which to split MMC design matrices between
paediatric and non-paediatric populations, the former of which are constant
over time. Set to NULL if not desired, Default: NULL}

\item{inc_time_tmc}{Indicator variable which decides whether to include
temporal random effects for TMC as well as MMC, Default: FALSE}
}
\value{
List of indicators of length 4.
}
\description{
Create indicators for the TMB model. These include:
\itemize{
\item{"is_type"}{Whether type information is available for the country in
question, as indicated by the columns \code{obs_mmc} and \code{obs_tmc} in \code{out}.}
\item{"rw_order"}{Whether an AR 1 (\code{FALSE}) or a RW x temporal prior should
be used (\code{TRUE}).}
\item{"paed_age_cutoff"}{Whether a paediatric MMC age cutoff should be
incorporated into the model.}
\item{inc_time_tmc}{Whether we want to include a time effect for TMC.}
}
}
\keyword{internal}
