% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrices.R
\name{create_design_matrices}
\alias{create_design_matrices}
\title{Create Design Matrices}
\usage{
create_design_matrices(dat, area_lev, k_dt = 5)
}
\arguments{
\item{dat}{Shell dataset (datputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{area_lev}{Desired admin boundary level to perform the analysis on.}

\item{k_dt}{Age knot spacing in spline definitions, Default: 5}
}
\value{
List of design matrices for fixed and random effects for medical
and traditional circumcision.
}
\description{
Create design matrices for fixed effects and temporal, age,
spatial and random effects, for both medical and traditional circumcision.
}
\seealso{
\code{\link[threemc]{create_shell_dataset}}
\code{\link[splines]{splineDesign}}
\code{\link[mgcv]{tensor.prod.model.matrix}}
\code{\link[methods]{as}}
\code{\link[Matrix]{sparse.model.matrix}}
}
