% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs_r.R
\name{create_dirs_r}
\alias{create_dirs_r}
\title{Recursively Create Missing Directories}
\usage{
create_dirs_r(dir_path)
}
\arguments{
\item{dir_path}{Path to a file or directory which you want to generate.}
}
\description{
Function to recursively create directories if any of the
directories in a provided path are missing. Somewhat equivalent to
\code{mkdir -p} in bash.
}
