% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hazard_matrix_agetime.R
\name{create_hazard_matrix_agetime}
\alias{create_hazard_matrix_agetime}
\title{Create Matrix Selecting Instantaneous Hazard Rate}
\usage{
create_hazard_matrix_agetime(
  dat,
  areas,
  area_lev,
  subset = NULL,
  time1 = "time1",
  time2 = "time2",
  timecaps = c(1, Inf),
  Ntime = NULL,
  age = "age",
  Nage = NULL,
  strat = NULL,
  Nstrat = NULL,
  circ = "circ",
  aggregated = FALSE,
  weight = NULL
)
}
\arguments{
\item{dat}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{PSNU area level for specific country.}

\item{subset}{Subset for dataset, Default: NULL}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored,
Default: "time2"}

\item{timecaps}{Window to fix temporal dimension before and after,
Default: c(1, Inf)}

\item{Ntime}{Number of time points (if NULL, function will calculate),
Default: NULL}

\item{age}{\itemize{
\item Variable with age circumcised or censored. Default: "age"
}}

\item{Nage}{Number of age groups (if NULL, function will calculate),
Default: NULL}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: NULL}

\item{Nstrat}{Number of stratification groups (if NULL, function will
calculate), Default: NULL}

\item{circ}{Variables with circumcision matrix, Default: "circ"}
}
\value{
Matrix for selecting instantaneous hazard rate.
}
\description{
Create a matrix selecting the instantaneous hazard rate needed
for survival analysis by age and time. The option to include an additional
stratification variable is also available, creating a 3D hazard function.
}
\seealso{
\code{\link[threemc]{create_shell_dataset}}
}
