% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integration_matrices.R
\name{create_integration_matrices}
\alias{create_integration_matrices}
\title{Create Integration Matrices for Selecting Instantaneous Hazard}
\usage{
create_integration_matrices(
  out,
  area_lev,
  time1 = "time1",
  time2 = "time2",
  age = "age",
  strat = "space",
  ...
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{area_lev}{PSNU area level for specific country. Defaults to the
Defaults to the maximum area level found in \code{dat} if not supplied.}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored,
Default: "time2"}

\item{age}{\itemize{
\item Variable with age circumcised or censored. Default: "age"
}}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: NULL}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{list} of length 2 of integration matrices for selecting
instantaneous hazard rate.
}
\description{
Create (unlagged and lagged) integration matrices for selecting
instantaneous hazard rate, required for fitting TMB model.
}
\seealso{
\code{\link[threemc]{create_shell_dataset}}
\code{\link[threemc]{create_integration_matrix_agetime}}
\code{\link[threemc]{create_integration_matrix_agetime_lag}}
}
