% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integration_matrix_agetime.R
\name{create_integration_matrix_agetime}
\alias{create_integration_matrix_agetime}
\title{Create Matri, simplify = FALSEx to Estimate Cumulative Hazard Rate}
\usage{
create_integration_matrix_agetime(
  dat,
  subset = NULL,
  time1 = "time1",
  time2 = "time2",
  timecaps = c(1, Inf),
  Ntime = NULL,
  age = "age",
  Nage = NULL,
  strat = NULL,
  Nstrat = NULL
)
}
\arguments{
\item{dat}{Dataset used for modelling.}

\item{subset}{Subset for dataset.}

\item{time1}{Variable name for time of birth.}

\item{time2}{Variable name for time circumcised or censored.}

\item{timecaps}{Window to fix temporal dimension before and after.}

\item{Ntime}{Number of time points (if NULL, function will calculate).}

\item{age}{\itemize{
\item Variable with age circumcised or censored.
}}

\item{Nage}{Number of age groups (if NULL, function will calculate).}

\item{strat}{Variable to stratify by in using a 3D hazard function.}

\item{Nstrat}{Number of stratification groups (if NULL, function will
calculate).}
}
\value{
Matrix for selecting instantaneous hazard rate.
}
\description{
Create a matrix to estimate the cumulative hazard rate needed
for survival analysis by age and time. The option to include an additional
stratification variable is also available, creating a 3D hazard function.
}
