% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integration_matrix_agetime_lag.R
\name{create_integration_matrix_agetime_lag}
\alias{create_integration_matrix_agetime_lag}
\title{Create Matrix to Estimate Lagged Cumulative Hazard Rate}
\usage{
create_integration_matrix_agetime_lag(
  dat,
  subset = NULL,
  time1 = "time1",
  time2 = "time2",
  timecaps = c(1, Inf),
  Ntime = NULL,
  age = "age",
  Nage = NULL,
  strat = NULL,
  Nstrat = NULL
)
}
\arguments{
\item{dat}{Dataset used for modelling.}

\item{subset}{Subset for dataset, Default: NULL}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored, Default: "time2"}

\item{timecaps}{Window to fix temporal dimension before and after,
Default: c(1, Inf)}

\item{Ntime}{Number of time points (if NULL, function will calculate),
Default: NULL}

\item{age}{\itemize{
\item Variable with age circumcised or censored, Default: "age"
}}

\item{Nage}{Number of age groups (if NULL, function will calculate),
Default: NULL}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: NULL}

\item{Nstrat}{Number of stratification groups (if NULL, function will
calculate), Default: NULL}
}
\value{
Matrix for selecting instantaneous hazard rate.
}
\description{
Create a matrix to estimate the lagged cumulative hazard rate
needed for survival analysis by age and time. The option to include an
additional stratification variable is also available, creating a 3D hazard
function.
}
\seealso{
\code{\link[Matrix]{sparseMatrix}}
}
