% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rw_prec_matrix.R
\name{create_rw_prec_matrix}
\alias{create_rw_prec_matrix}
\title{Create Precision Matrix for RWp Process}
\usage{
create_rw_prec_matrix(dim, order = 1, offset.diag = TRUE)
}
\arguments{
\item{dim}{Dimension of the precision matrix.}

\item{order}{Order of the random walk, Default: 1}

\item{offset.diag}{Option to offset diagonal by 1E-6, Default: TRUE}
}
\value{
RW precision matrix
}
\description{
Create the precision matrix for a RWp process.
}
