% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survival_matrices.R
\name{create_survival_matrices}
\alias{create_survival_matrices}
\title{Create Survival Matrices for Selecting Instantaneous Hazard}
\usage{
create_survival_matrices(
  out,
  time1 = "time1",
  time2 = "time2",
  age = "age",
  strat = "space",
  ...
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored,
Default: "time2"}

\item{age}{\itemize{
\item Variable with age circumcised or censored. Default: "age"
}}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: "space"}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{list} of length 4 of survival matrices for selecting
instantaneous hazard rate.
}
\description{
Create empirical agetime hazard matrices for medical and
traditional circumcision, as well as for censored (i.e. non-circumcised) and
left-censored (i.e. age of circumcision unknown) individuals.
}
\seealso{
\code{\link[threemc]{create_shell_dataset}}
\code{\link[threemc]{create_hazard_matrix_agetime}}
}
