% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapack_psnu_area_level}
\alias{datapack_psnu_area_level}
\title{PSNU Area Levels for SSA}
\format{
A \code{data.frame} with 29 rows and 2 variables:
}
\usage{
data(datapack_psnu_area_level)
}
\description{
PSNU area levels for Sub-Saharan African countries. These
are the recommended levels at which to perform modelling etc., for each
respective country. Inferences on larger regions (i.e. lower PSNU area
levels) can be made by aggregating results for higher area levels. The
dataset contains the following fields:
\itemize{
\item{\code{iso3}}{character ISO3 codes for various Sub-Saharan African
countries.}
\item{\code{psnu_area_level}}{integer The sub national level considered
to be the organizational level in which a country has prioritised their
program. Increasing values refer to more granular regional distinctions.}
}
}
\keyword{datasets}
