% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_areas}
\alias{demo_areas}
\title{Malawi shapefiles}
\format{
A \code{sf} collecton of 6 features and 9 fields, including a
\code{data.frame} with 387 rows, 10 variables, and a sf
}
\usage{
data(demo_areas)
}
\description{
\code{sf} shapefile representation of Malawi, as a multipolygon.
\itemize{
\item{\code{iso3}}{character ISO3 codes for Sub-Saharan African
countries.}
\item{\code{area_id}}{Unique ID for each region in MWI. Formatted as
"County_area_level_ID" (e.g. MWI_3_05 for Mzimba)}
\item{\code{area_name}}{Name of region in question}
\item{\code{parent_area_id} {Unique ID for region's parent region.}}
\item{\code{area_level} {Numeric value denoting area level of
area, in decreasing granularity.}}
\item{\code{area_level_label} {Translates numeric area level to meaning
in country in question. For example, in Malawi a region of area level 3
is a "District".}}
\item{\code{area_sort_order} {Order to sort areas in when plotting,
roughly equivalent to a \code{geofacet} grid.}}
\item{\code{center_x} {X coordinate for centre of region's multipolygon}}
\item{\code{center_y} {Y coordinate for centre of region's multipolygon}}
\item{\code{geometry}} {\code{sfc_MULTIPOLYGON} representation of region's
spatial geometry}
}
}
\keyword{datasets}
