% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_populations}
\alias{demo_populations}
\title{Malawi populations}
\format{
A \code{data.frame} with 58806 rows and 7 variables.
}
\usage{
data(demo_populations)
}
\description{
Single age, aggregated male populations for each area in
Malawi.
\itemize{
\item{\code{iso3}}{character ISO3 codes for Sub-Saharan African
countries.}
\item{\code{area_id}}{Unique ID for each region in MWI. Formatted as
"County_area_level_ID" (e.g. MWI_3_05 for Mzimba)}
\item{\code{area_level} {Numeric value denoting area level of
area, in decreasing granularity.}}
\item{\code{area_name}}{Name of region in question}
\item{\code{year}}{Year for population in question}
\item{\code{age}}{Age for population in question}
\item{\code{population} {(Male) Population for each unique
area-year-age combination.}}
}
}
\keyword{datasets}
