% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_survey_circumcision}
\alias{demo_survey_circumcision}
\title{Malawi surveys}
\format{
A \code{data.frame} with 29313 rows and 12 variables.
}
\usage{
data(demo_survey_circumcision)
}
\description{
Circumcision surveys for Malawi.
\itemize{
\item{\code{iso3}}{character ISO3 codes for Sub-Saharan African
countries.}
\item{\code{survey_id}}{Survey id for each record.}
\item{\code{area_id}}{Unique ID for each region in MWI. Formatted as
"County_area_level_ID" (e.g. MWI_3_05 for Mzimba)}
\item{\code{area_level} {Numeric value denoting area level of
area, in decreasing granularity.}}
\item{\code{age}{Age at interview.}}
\item{\code{dob_cmc}{CMC (Century Month Code) date of birth of 
  individual.)}}
\item{\code{interview_cmc}{CMC date of interview.}}
\item{\code{indweight}{Weighting for survey record in question}}
\item{\code{circ_status}{Circumcision status of individual, 1 indicating 
  circumcision and 0 indicating right-censoring.}}
\item{\code{circ_age}{Age at circumcision, if applicable.}}
\item{\code{circ_who}{Circumcision provider, either medical or 
  traditional.}}
\item{\code{circ_where}{Circumcision location, either medical or 
  traditional.}}
}
}
\keyword{datasets}
