% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esa_wca_regions}
\alias{esa_wca_regions}
\title{WCA - ESA key for Sub-Saharan African countries}
\format{
A \code{data.frame} with 38 rows and 3 variables:
}
\usage{
data(esa_wca_regions)
}
\description{
Western and Central Africa (WCA) - Eastern and Southern Africa
(ESA) categorisation for Sub-Saharan African countries. Also includes
North-South-East-West categorisation.
\itemize{
\item{\code{iso3}}{character ISO3 codes for Sub-Saharan African
countries.}
\item{\code{region}}{character ESA-WCA categorisation for each
\code{iso3}}
\item{\code{four_region}}{character North-South-East-West categorisation for
each \code{iso3}}
}
}
\keyword{datasets}
