% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fill_downup_populations}
\alias{fill_downup_populations}
\title{Assume constant historical populations}
\usage{
fill_downup_populations(populations, start_year, min_pop_year = NULL)
}
\description{
Fills populations for historical years not present in
populations dataset with earliest known population for each unique area_id -
age combination.
}
\keyword{internal}
