% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{match_age_group_to_ages}
\alias{match_age_group_to_ages}
\title{Create data frame of all ages within provided age group.}
\usage{
match_age_group_to_ages(age_group, max_age = 60)
}
\arguments{
\item{age_group}{Age group, either "x-x" for a fixed upper age, or "x+", for
an age group with an upper age of \code{max_age}.}

\item{max_age}{Maximum age for age groups with no upper limit, Default: 60}
}
\description{
Create data frame of all ages within provided age group.
}
\keyword{internal}
