% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_functions.R, R/merge_area_info.R
\name{merge_area_info}
\alias{merge_area_info}
\title{Merge Regional Informatoin on Dataset}
\usage{
merge_area_info(results, areas)

merge_area_info(results, areas)
}
\arguments{
\item{results}{\code{data.frame} you wish to merge shapefiles with.}

\item{areas}{\code{sf} shapefiles for specific country/region.}
}
\description{
Merge regional information on the dataset
(i.e. parent area info).

Merge regional information on the dataset
(i.e. parent area info).
}
\keyword{internal}
