% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{n_circumcised}
\alias{n_circumcised}
\title{Calculate Number of People Circumcised}
\usage{
n_circumcised(results)
}
\arguments{
\item{results}{Results with samples for number of circumcisions performed
in each region.}
}
\description{
Calculate number of people circumcised (as well as unmet need).
}
\keyword{internal}
