% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_survey_data.R
\name{normalise_weights_kish}
\alias{normalise_weights_kish}
\title{Normalise Survey Weights and apply Kish Coefficients}
\usage{
normalise_weights_kish(
  survey_circumcision,
  strata.norm = c("survey_id", "area_id"),
  strata.kish = c("survey_id")
)
}
\arguments{
\item{survey_circumcision}{Information on male circumcision status from
surveys containing survey weights.}

\item{strata.norm}{Stratification variables for normalising survey weights,
Default: c("survey_id", "area_id")}

\item{strata.kish}{Stratification variables for estimating and applying the
Kish coefficients, Default: "survey_id"}
}
\value{
Survey data with normalised survey weights and required variables to
run circumcision model.
}
\description{
Normalise survey weights and apply Kish coefficients.
}
\keyword{internal}
