% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{patterns}
\alias{patterns}
\title{Pattern matching for \code{data.table}}
\usage{
patterns(..., cols = character(0L))
}
\description{
patterns returns the matching indices in the argument \code{cols}
corresponding to the regular expression patterns provided. The patterns must
be supported by \code{grep.} See also \code{data.table:::patterns()}.
}
\keyword{internal}
