% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{prepare_sample_data}
\alias{prepare_sample_data}
\title{Pull N circumcision samples from TMB fit}
\usage{
prepare_sample_data(
  N = 100,
  populations,
  no_prog_results = NULL,
  prog_results = NULL,
  no_prog_tmb_fit,
  prog_tmb_fit,
  type
)
}
\arguments{
\item{N}{Number of samples to be generated, Default: 100}

\item{populations}{\code{data.frame} containing populations for each
region in tmb fits.}

\item{no_prog_results}{Quantiles of different summaries for different types
(Medical, traditional or total) circumcision, for survey data only,
Default: NULL}

\item{prog_results}{Quantiles of different summaries for different types
(Medical, traditional or total) circumcision, for survey data and VMMC
programme data, Default: NULL}

\item{no_prog_tmb_fit}{TMB model for Male Circumcision (MC), for survey data
only.}

\item{prog_tmb_fit}{TMB model for Male Circumcision (MC), for survey data and
VMMC programme data.}

\item{type}{Determines which aspect of MC in the regions in question you wish
to sample for. Can be one of "probability", "incidence" or "prevalence".}
}
\description{
Function to pull samples for various summaries inferred about
circumcision in a given area (prevalence, probability or incidence,
respectively).
}
\keyword{internal}
