% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_survey_aggregation.R
\name{prepare_survey_aggregation}
\alias{prepare_survey_aggregation}
\title{Format circumcision survey data with age data reformatted.}
\usage{
prepare_survey_aggregation(
  areas_wide,
  survey_circumcision,
  survey_individuals,
  survey_clusters
)
}
\arguments{
\item{areas_wide}{\code{data.frame} with shapefiles and area hierarchy.}

\item{survey_circumcision}{\itemize{
\item Information on male circumcision status from
surveys.
}}

\item{survey_individuals}{\itemize{
\item Information on the individuals surveyed.
}}

\item{survey_clusters}{\itemize{
\item Information on the survey clusters.
}}
}
\value{
Survey data, with age data reformatted, at highest/most granular
area level.
}
\description{
Format circumcision survey data with age data reformatted.
}
