% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_survey_data.R
\name{prepare_survey_data}
\alias{prepare_survey_data}
\title{Prepare Survey Data}
\usage{
prepare_survey_data(
  areas,
  survey_circumcision,
  survey_individuals,
  survey_clusters,
  area_lev,
  start_year = 2006,
  cens_year = NULL,
  cens_age = 59,
  rm_missing_type = FALSE,
  norm_kisk_weights = TRUE,
  strata.norm = c("survey_id", "area_id"),
  strata.kish = c("survey_id")
)
}
\arguments{
\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{survey_circumcision}{\itemize{
\item Information on male circumcision status from
surveys.
}}

\item{survey_individuals}{\itemize{
\item Information on the individuals surveyed.
}}

\item{survey_clusters}{\itemize{
\item Information on the survey clusters.
}}

\item{area_lev}{\itemize{
\item Desired admin boundary level to perform the analysis on.
}}

\item{start_year}{\itemize{
\item Year to begin the analysis on, Default: 2006
}}

\item{cens_year}{\itemize{
\item Year to censor the circumcision data by (Sometimes some
weirdness at the final survey year, e.g. v small number of MCs),
Default: NULL
}}

\item{cens_age}{\itemize{
\item Age to censor the circumcision data at, Default: 59
}}

\item{rm_missing_type}{\itemize{
\item Indicator to decide whether you would like to keep
surveys where there is no MMC/TMC disinction. These surveys may still be
useful for determining MC levels, Default: FALSE
}}

\item{norm_kisk_weights}{\itemize{
\item Indicator to decide whether to normalise survey
weights and apply Kish coefficients, Default: TRUE
}}

\item{strata.norm}{Stratification variables for normalising survey weights,
Default: c("survey_id", "area_id")}

\item{strata.kish}{Stratification variables for estimating and applying the
Kish coefficients, Default: "survey_id"}
}
\value{
Survey data with required variables to run circumcision model.
}
\description{
Prepare survey data required to run the circumcision model. Can
also optionally apply \link[threemc]{normalise_weights_kish}, to
normalise survey weights and apply Kish coefficients.
}
\seealso{
\code{\link[threemc]{normalise_weights_kish}}
}
