% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{prevalence_change}
\alias{prevalence_change}
\title{Calculate Change in Prevalence/Coverage from a given year.}
\usage{
prevalence_change(results, spec_year)
}
\arguments{
\item{results}{\code{results} results for several years.}

\item{spec_year}{Year to calculate change in prevalence/coverage from within
\code{results}.}
}
\description{
Function to calculate change in prevalence/coverage from a
given year for all other years.
}
\keyword{internal}
