% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R, R/prepare_survey_data.R
\name{compare_predictions}
\alias{compare_predictions}
\alias{find_circ_type}
\alias{reassign_survey_level}
\title{Define circumcision type}
\usage{
compare_predictions(actual, predictions)

find_circ_type(survey_circumcision)

reassign_survey_level(
  survey_circumcision,
  areas,
  area_lev,
  remove_add_cols = TRUE
)
}
\arguments{
\item{actual}{\code{vector} of observed values.}

\item{predictions}{\code{dataframe/tibble/data.table} or \code{matrix} with
a row for each value in \code{actual}, and a column for each prediction.}

\item{survey_circumcision}{\itemize{
\item Information on male circumcision status from
surveys. If this is a list or contains more than one country, the function
is performed for each country present, returning a list.
}}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{\itemize{
\item Desired admin boundary level to perform the analysis on.
}}

\item{remove_add_cols}{Only keep columns originally present in
\code{survey_circumcision}, Default = TRUE.}
}
\value{
List of:
\itemize{
\item{"summary_stats_df"}{pointwise prediction stats as a \code{data.frame}}
\item{"summary_stats"}{mean/summed overall prediction stats}
}

Surveys with circumcision type labelled by the column \code{type}.

Surveys with records reassigned to \code{area_lev}.
}
\description{
Using \code{circ_who} and \code{circ_where}, determines survey
type.

Using \code{circ_who} and \code{circ_where}, determines survey
type.

Join surveys with areas for each level leading up to the
desired area level. This distributes granular surveys to desired area_level
using parent_ids.
}
\keyword{internal}
