% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rmvnorm_sparseprec}
\alias{rmvnorm_sparseprec}
\title{Take multivariate normal sample from sparse precision matrix}
\usage{
rmvnorm_sparseprec(n, mean = rep(0, nrow(prec)), prec = diag(length(mean)))
}
\description{
See also \code{naomi:::rmvnorm_sparseprec()}.
}
\keyword{internal}
