% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{scale_gmrf_precision}
\alias{scale_gmrf_precision}
\title{Scale of GMRF precision matrix}
\usage{
scale_gmrf_precision(
  Q,
  A = matrix(1, ncol = ncol(Q)),
  eps = sqrt(.Machine$double.eps)
)
}
\description{
See also \code{naomi::scale_gmrf_precision()}.
}
\keyword{internal}
