% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{shell_data_spec_area}
\alias{shell_data_spec_area}
\title{Subset and Prepare Shell Dataset to Only One Admin Boundary}
\usage{
shell_data_spec_area(dat, area_lev = NULL)
}
\arguments{
\item{dat}{Shell dataset used for modelling}

\item{area_lev}{PSNU area level for specific country. Defaults to the
Defaults to the maximum area level found in \code{dat} if not supplied.}
}
\description{
Subset the shell dataset to a specific area of interest and
resets the space counter. The output dataset can be used to set up model
components on the specified administrative boundaries.
}
\keyword{internal}
