% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{split_mmc_design_matrices_paed}
\alias{split_mmc_design_matrices_paed}
\title{Split MMC Design Matrices between adult and paediatric populations}
\usage{
split_mmc_design_matrices_paed(
  out,
  area_lev,
  design_matrices,
  paed_age_cutoff = 10
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{area_lev}{PSNU area level for specific country.}

\item{design_matrices}{Design matrices for fixed effects and temporal, age,
spatial and random effects, for both medical and traditional circumcision.}

\item{paed_age_cutoff}{Age at which to split MMC design matrices between
paediatric and non-paediatric populations, the former of which are constant
over time. Set to NULL if not desired, Default: NULL}
}
\description{
Take MMC-related design matrices and effectively "split"
them into two parts; one paediatric set of design matrices which does not
include any time-related components, and one non-paediatric set.
}
\keyword{internal}
