% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{survey_points_dmppt2_convert_convention}
\alias{survey_points_dmppt2_convert_convention}
\title{Create data frame of all ages within provided age group.
Convert survey coverage points & dmppt2 data to match convention of
aggregated results.}
\usage{
survey_points_dmppt2_convert_convention(.data)
}
\arguments{
\item{.data}{Data frame with either survey calculated coverage, with
associated error bounds, or DMPPT2 coverage estimates calculated from VMMC
programme data.}
}
\description{
Create data frame of all ages within provided age group.
Convert survey coverage points & dmppt2 data to match convention of
aggregated results.
}
