% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_functions.R
\name{threemc_aggregate}
\alias{threemc_aggregate}
\title{Produce Population weighted Aggregated Samples}
\usage{
threemc_aggregate(
  .data,
  fit,
  areas,
  populations,
  age_var = c("age", "age_group"),
  type = c("probability", "incidence", "prevalence"),
  area_lev,
  N = 100,
  prev_year = 2008
)
}
\arguments{
\item{.data}{\code{data.frame} of unaggregated modelling results.}

\item{fit}{\code{TMB} list containing model parameters, nested list of
\code{samples} for the (cumulative) incidence and hazard rate of circumcision for
the region(s) in question.}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{populations}{\code{data.frame} containing populations for each
region in tmb fits.}

\item{age_var}{Determines whether you wish to aggregate by discrete ages or
age groups (0-4, 5-9, 10-14, and so on).}

\item{type}{Determines which aspect of MC in the regions in question you wish
to aggregate for. Can be one of "probability", "incidence" or "prevalence".}

\item{area_lev}{PSNU area level for specific country.}

\item{N}{Number of samples to be generated, Default: 100}

\item{prev_year}{If type == "prevalence", choose year to compare prevalence
with.}
}
\value{
\code{data.frame} with samples aggregated by \code{aggr_cols} and
weighted by population.
}
\description{
Aggregate by area, year, age and type (weighted by population),
and convert to a percentage/probability.
}
