% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{threemc_fit_model}
\alias{threemc_fit_model}
\title{Produce TMB model fit with sample, or re-sample from existing
optimised model fit.}
\usage{
threemc_fit_model(
  fit = NULL,
  dat_tmb = NULL,
  mod,
  parameters = NULL,
  maps = NULL,
  randoms = c("u_time_mmc", "u_age_mmc", "u_space_mmc", "u_agetime_mmc",
    "u_agespace_mmc", "u_spacetime_mmc", "u_age_tmc", "u_space_tmc", "u_agespace_tmc"),
  sample = TRUE,
  smaller_fit_obj = FALSE,
  sdreport = FALSE,
  N = 1000,
  ...
)
}
\arguments{
\item{fit}{Optional "small" fit object with no \code{sample}. Specifying \code{fit}
means you do not need to specify \code{dat_tmb} or \code{parameters}, as argument
specifications will be overridden by those stored in \code{fit}.}

\item{dat_tmb}{\code{list} of data required for model fitting, outputted
by \link[threemc]{threemc_prepare_model_data}, which includes:
\itemize{
\item{design_matrices}{Includes\code{X_fixed_mmc}, \code{X_fixed_tmc}, \code{X_time_mmc},
\code{X_age_mmc}, \code{X_age_tmc}, \code{X_space_mmc}, \code{X_space_tmc}, \code{X_agetime_mmc},
\code{X_agespace_mmc}, \code{X_agespace_tmc}, \code{X_spacetime_mmc}. Design
Create design matrices for fixed effects and temporal, age, space and
interaction random effects}
\item{integration matrices}{Includes \code{IntMat1}, \code{IntMat2}. Integration
matrices for selecting the instantaneous hazard rate.}
\item{survival matrices}{Includes \code{A_mmc}, \code{A_tmc}, \code{A_mc}, \code{B}, \code{C}.
Survival matrices for MMC, TMC, censored and left censored}
\item{Q_space}{Precision/Adjacency matrix for the spatial random effects.
}
}}

\item{mod}{TMB model, one of either
"Surv_SpaceAgeTime_ByType_withUnknownType" or "Surv_SpaceAgeTime" if the
surveys for the country in question make no distinction between circumcision
type (i.e whether they were performed in a medical or traditional setting).}

\item{parameters}{\code{list} of fixed and random model parameters.}

\item{maps}{\code{list} of factors with value NA, the names of which
indicate parameters to be kept fixed at their initial value throughout the
optimisation process.}

\item{randoms}{\code{vector} of random effects.}

\item{sample}{If set to TRUE, has function also return N samples for
medical, traditional and total circumcisions, Default: TRUE}

\item{smaller_fit_obj}{Returns a smaller fit object. Useful for saving the
fit object for later aggregations.}

\item{sdreport}{If set to TRUE, produces the standard deviation report for
the model, Default: FALSE}

\item{N}{Number of samples to be generated, Default: 1000}

\item{...}{Further arguments passed to internal functions.}
}
\value{
TMB model fit, including optimised parameters, hessian matrix,
samples and standard deviation report (if desired).
}
\description{
Optimises threemc objective function and produces samples from
model fit (if so desired). If provided with an existing optimised model
\code{fit}, can also perform re-sampling.
}
