% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{threemc_initial_pars}
\alias{threemc_initial_pars}
\title{Initialise \code{thremec} (hyper)parameters.}
\usage{
threemc_initial_pars(
  dat_tmb,
  custom_init = NULL,
  rw_order = NULL,
  paed_age_cutoff = NULL,
  inc_time_tmc = FALSE
)
}
\arguments{
\item{dat_tmb}{\code{list} of data required for model fitting, outputted
by \link[threemc]{threemc_prepare_model_data}, which includes:
\itemize{
\item{design_matrices}{Includes\code{X_fixed_mmc}, \code{X_fixed_tmc}, \code{X_time_mmc},
\code{X_age_mmc}, \code{X_age_tmc}, \code{X_space_mmc}, \code{X_space_tmc}, \code{X_agetime_mmc},
\code{X_agespace_mmc}, \code{X_agespace_tmc}, \code{X_spacetime_mmc}. Design
Create design matrices for fixed effects and temporal, age, space and
interaction random effects}
\item{integration matrices}{Includes \code{IntMat1}, \code{IntMat2}. Integration
matrices for selecting the instantaneous hazard rate.}
\item{survival matrices}{Includes \code{A_mmc}, \code{A_tmc}, \code{A_mc}, \code{B}, \code{C}.
Survival matrices for MMC, TMC, censored and left censored}
\item{Q_space}{Precision/Adjacency matrix for the spatial random effects.
}
}}

\item{custom_init}{named \code{list} of custom fixed and random
model parameters you want to supersede "hardcoded" defaults, default = NULL.}

\item{rw_order}{Order of the random walk used for temporal precision matrix.
Setting to NULL assumes you wish to specify an AR 1 temporal prior.
Default: NULL}

\item{paed_age_cutoff}{Age at which to split MMC design matrices between
paediatric and non-paediatric populations, the former of which are constant
over time. Set to NULL if not desired, Default: NULL}

\item{inc_time_tmc}{Indicator variable which decides whether to include
temporal random effects for TMC as well as MMC, Default: FALSE}
}
\value{
Named \code{list} of intial (hyper)parameters for
\code{threemc_fit_model}
}
\description{
Return minimised fit object. Often useful when saving the fit
object for later aggregation.
}
