% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_2.R
\name{threemc_ppc2}
\alias{threemc_ppc2}
\title{Posterior Predictive Distribution and checks on OOS survey}
\usage{
threemc_ppc2(
  fit,
  out,
  survey_circumcision_test,
  areas = NULL,
  area_lev = 1,
  age_groups = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "54-59"),
  N = 1000,
  seed = 123
)
}
\arguments{
\item{fit}{Fit object returned by \code{naomi::sample_tmb}, which includes,
among other things, the optimised parameters and subsequent sample for our
TMB model.}

\item{out}{Results of model fitting (at specified model \code{area_lev})
outputted by \link[threemc]{compute_quantiles}.}

\item{survey_circumcision_test}{\code{survey_circumcision} dataset loaded with
\link[threemc]{read_circ_data}. \emph{Do not preprocess with}
\emph{\link[threemc]{prepare_survey_data}} If performing an OOS validation of
model performance, you should filter this dataset for the years "held back"
from your model fit.}

\item{areas}{\code{sf} shapefile for specific country/region. Only required
if \code{survey_circumcision_test} has records for area levels higher (i.e. more
granular) than \code{area_lev}, in which case they must be reassigned to their
\code{parent_area_id} at \code{area_lev}, Default = NULL.}

\item{area_lev}{Area level you wish to aggregate to when performing posterior
predictive comparisons with survey estimates.}

\item{age_groups}{Age groups to aggregate by, Default:
c("0-4",   "5-9",   "10-14", "15-19", "20-24", "25-29",
"30-34", "35-39", "40-44", "45-49", "50-54", "54-59")}

\item{N}{Number of samples to generate, Default: 1000}

\item{seed}{Random seed used for taking binomial sample from posterior
predictive distribution.}

\item{type}{Decides type of circumcision coverage to perform PPC on, must
be one of "MC", "MMC", or "TMC", Default = "MMC"}

\item{CI_range}{CI interval about which you want to compare empirical and
posterior predictive estimates for left out surveys,
Default = c(0.5, 0.8, 0.95)}
}
\value{
\code{data.frame} with samples aggregated by \code{aggr_cols} and
weighted by population.
}
\description{
Aggregate specified \code{numeric} columns by population-weighted
age groups (rather than single year ages), split by specified categories.
Using an alternative method to previously.
}
