% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{threemc_prepare_model_data}
\alias{threemc_prepare_model_data}
\title{Produce Data Matrices for Modelling}
\usage{
threemc_prepare_model_data(
  out,
  areas,
  area_lev = NULL,
  aggregated = TRUE,
  weight = "population",
  k_dt_age = 5,
  k_dt_time = 5,
  paed_age_cutoff = NULL,
  rw_order = NULL,
  inc_time_tmc = FALSE,
  ...
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{PSNU area level for specific country.}

\item{aggregated}{\code{agggregated = FALSE} treats every area_id as its own
object, allowing for the use of surveys for lower area hierarchies.
\code{aggregated = TRUE} means we only look at area level of interest.}

\item{weight}{variable to weigh circumcisions by when aggregating for
lower area hierarchies (only applicable for \code{aggregated = TRUE})}

\item{k_dt_age}{Age knot spacing in spline definitions, Default: 5}

\item{k_dt_time}{Time knot spacing in spline definitions, set to NULL to
disable temporal splines, Default: 5}

\item{paed_age_cutoff}{Age at which to split MMC design matrices between
paediatric and non-paediatric populations, the former of which are constant
over time. Set to NULL if not desired, Default: NULL}

\item{rw_order}{Order of the random walk used for temporal precision matrix.
Setting to NULL assumes you wish to specify an AR 1 temporal prior.
Default: NULL}

\item{inc_time_tmc}{Indicator variable which decides whether to include
temporal random effects for TMC as well as MMC, Default: FALSE}

\item{...}{Additional arguments to be passed to functions which create
matrices.}
}
\value{
\code{list} of data required for model fitting, including:
\itemize{
\item{design_matrices}{Includes\code{X_fixed_mmc}, \code{X_fixed_tmc}, \code{X_time_mmc},
\code{X_age_mmc}, \code{X_age_tmc}, \code{X_space_mmc}, \code{X_space_tmc}, \code{X_agetime_mmc},
\code{X_agespace_mmc}, \code{X_agespace_tmc}, \code{X_spacetime_mmc}. Design
Create design matrices for fixed effects and temporal, age, space and
interaction random effects}
\item{integration matrices}{Includes \code{IntMat1}, \code{IntMat2}. Integration
matrices for selecting the instantaneous hazard rate.}
\item{survival matrices}{Includes \code{A_mmc}, \code{A_tmc}, \code{A_mc}, \code{B}, \code{C}.
Survival matrices for MMC, TMC, censored and left censored}
\item{Q_space}{Precision/Adjacency matrix for the spatial random effects.
}
}
}
\description{
Create data for modelling. Output detailed below.
}
