% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{threemc_resample}
\alias{threemc_resample}
\title{Resample from minimal TMB fit}
\usage{
threemc_resample(fit, obj, N = 1000, sdreport = FALSE)
}
\arguments{
\item{fit}{Optional "small" fit object with no \code{sample}. Specifying \code{fit}
means you do not need to specify \code{dat_tmb} or \code{parameters}, as argument
specifications will be overridden by those stored in \code{fit}.}

\item{obj}{TMB object/AD model outputted by \link[TMB]{MakeADFun}.}

\item{N}{Number of samples to be generated, Default: 1000}

\item{sdreport}{If set to TRUE, produces the standard deviation report for
the model, Default: FALSE}
}
\value{
Object of class "naomi_fit", containing the original TMB object
("obj"), the standard deviation report for optimised AD model (from
\link[TMB]{sdreport}) and \code{n_samples} samples for the (cumulative) incidence
and hazard rate of circumcision for the region(s) in question.
}
\description{
Sample from TMB object, using \code{naomi::sample_tmb}. Saves
changing object to \code{naomi} format. Also produces and returns standard
deviation report outputted by \link[TMB]{sdreport}, if desired.
}
\keyword{internal}
