% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_sample}
\alias{aggregate_sample}
\title{Produce Population Weighted Aggregated Samples}
\usage{
aggregate_sample(
  .data,
  aggr_cols = c("area_id", "area_name", "year", "age", "age_group", "model", "type"),
  num_cols,
  ...
)
}
\arguments{
\item{.data}{\code{data.frame} including area populations, with
un-aggregated samples.}

\item{aggr_cols}{Columns to aggregate samples by, Default:
c("area_id", "area_name", "year", "age", "age_group", "model", "type")}

\item{num_cols}{\code{numeric} columns to aggregate.}

\item{...}{Further arguments passed to \code{data.table::rbindlist}.}
}
\value{
\code{data.frame} with samples aggregated by \code{aggr_cols} and
weighted by population.
}
\description{
Aggregate by area, year, age and type (weighted by population),
and convert to a percentage/probability.
}
\keyword{internal}
