% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{circ_sample_tmb}
\alias{circ_sample_tmb}
\title{Sample TMB fit for Circumcision Model}
\usage{
circ_sample_tmb(
  fit = NULL,
  obj = NULL,
  opt,
  sdreport = FALSE,
  nsample = 1000,
  ...
)
}
\arguments{
\item{obj}{TMB object/AD model outputted by \link[TMB]{MakeADFun}.}

\item{opt}{Optimised TMB model, outputted by optimisation function such
as \link[stats]{nlminb} or \link[stats]{optim}.}

\item{nsample}{Number of samples to be generated, Default: 1000}

\item{...}{Further arguments passed to \link[naomi]{sample_tmb}.}
}
\value{
Object of class "naomi_fit", containing the original TMB object
("obj"), the standard deviation report for optimised AD model (from
\link[TMB]{sdreport}) and \code{n_samples} samples for the (cumulative) incidence
and hazard rate of circumcision for the region(s) in question.
}
\description{
Sample from TMB object, using \link[naomi]{sample_tmb}. Saves
changing object to "Naomi" format. Also produces and returns standard
deviation report outputted by \link[TMB]{sdreport}.
}
\seealso{
\code{\link[TMB]{sdreport}}
\code{\link[naomi]{sample_tmb}}
}
\keyword{internal}
