% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregate_structure.R
\name{create_aggregate_structure}
\alias{create_aggregate_structure}
\title{Create a List containing the Hierarchy of levels}
\usage{
create_aggregate_structure(areas, area_lev)
}
\arguments{
\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{PSNU area level for specific country.}
}
\description{
Create a list containing all the area dependencies and number
of for each area in the hierarchy
}
