% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_icar_prec_matrix.R
\name{create_icar_prec_matrix}
\alias{create_icar_prec_matrix}
\title{Create Precision Matrix for ICAR Process}
\usage{
create_icar_prec_matrix(sf_obj = NULL, row.names = NULL)
}
\arguments{
\item{sf_obj}{Shapefiles needed for adjacency, Default: NULL}

\item{row.names}{Unique IDs for the areas, Default: NULL}
}
\value{
ICAR precision matrix.
}
\description{
Create the precision matrix for an ICAR process.
}
\seealso{
\code{\link[spdep]{poly2nb}}
\code{\link[spdep]{nb2mat}}
\code{\link[naomi]{scale_gmrf_precision}}
\code{\link[methods]{as}}
}
