% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_dataset.R
\name{create_shell_dataset}
\alias{create_shell_dataset}
\title{Create Shell Dataset for Estimating Empirical Circumcision Rate}
\usage{
create_shell_dataset(
  survey_circumcision,
  populations,
  areas,
  area_lev = NULL,
  start_year,
  end_year = 2021,
  time1 = "time1",
  time2 = "time2",
  strat = "space",
  age = "age",
  circ = "indweight_st",
  ...
)
}
\arguments{
\item{survey_circumcision}{\itemize{
\item Information on male circumcision status from
surveys. If this is a list or contains more than one country, the function
is performed for each country present, returning a list.
}}

\item{populations}{\code{data.frame} containing populations for each
region in tmb fits.}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{\itemize{
\item Desired admin boundary level to perform the analysis on.
}}

\item{start_year}{First year in shell dataset.}

\item{end_year}{Last year in shell dataset, which is also the year to
forecast/model until, Default: 2021}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored,
Default: "time2"}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: NULL}

\item{age}{\itemize{
\item Variable with age circumcised or censored. Default: "age"
}}

\item{circ}{Variables with circumcision matrix, Default: "indweight_st"}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{data.frame} with a row for every unique record in
\code{survey_circumcision} for a given area. Also includes empirical estimates
for circumcision estimates for each unique record.
}
\description{
Create a shell dataset with a row for every unique area ID,
area name, year and circumcision age in survey data. Also, computes the
empirical number of person years until circumcision and number of people
circumcised for several "types" of circumcision; known medical
circumcisions, known traditional circumcisions, censored survey entries
(i.e. where surveyed individuals had not been circumcised) and left-censored
survey entries (i.e. where circumcision occurred at an unknown age).
}
\seealso{
\code{\link[threemc]{datapack_psnu_area_level}}
\code{\link[tidyr]{crossing}}
\code{\link[threemc]{create_integration_matrix_agetime}}
\code{\link[threemc]{create_hazard_matrix_agetime}}
}
