% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{create_survival_matrices}
\alias{create_survival_matrices}
\title{Create Survival Matrices for Selecting Instantaneous Hazard}
\usage{
create_survival_matrices(
  out,
  areas = areas,
  area_lev = area_lev,
  time1 = "time1",
  time2 = "time2",
  age = "age",
  strat = "space",
  aggregated = TRUE,
  weight = "population",
  ...
)
}
\arguments{
\item{out}{Shell dataset (outputted by \link[threemc]{create_shell_dataset}
with a row for every unique record in circumcision survey data for a given
area. Also includes empirical estimates for circumcision estimates for each
unique record.}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{PSNU area level for specific country.}

\item{time1}{Variable name for time of birth, Default: "time1"}

\item{time2}{Variable name for time circumcised or censored,
Default: "time2"}

\item{age}{\itemize{
\item Variable with age circumcised or censored. Default: "age"
}}

\item{strat}{Variable to stratify by in using a 3D hazard function,
Default: NULL}

\item{aggregated}{\code{agggregated = FALSE} treats every area_id as its own
object, allowing for the use of surveys for lower area hierarchies.
\code{aggregated = TRUE} means we only look at area level of interest.}

\item{weight}{variable to weigh circumcisions by when aggregating for
lower area hierarchies (only applicable for \code{aggregated = TRUE})}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{list} of length 4 of survival matrices for selecting
instantaneous hazard rate.
}
\description{
Create empirical agetime hazard matrices for medical and
traditional circumcision, as well as for censored (i.e. non-circumcised) and
left-censored (i.e. age of circumcision unknown) individuals.
}
\seealso{
\code{\link[threemc]{create_shell_dataset}}
\code{\link[threemc]{create_hazard_matrix_agetime}}
}
\keyword{internal}
