% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{minimise_fit_obj}
\alias{minimise_fit_obj}
\title{Minimise Fit Object Size}
\usage{
minimise_fit_obj(fit, dat_tmb, paramaters)
}
\arguments{
\item{fit}{Fit object returned by \link[naomi]{sample_tmb}, which includes,
among other things, the optimised parameters and subsequent sample for our
TMB model.}

\item{dat_tmb}{\code{list} of data required for model fitting, outputted
by \link[threemc]{threemc_prepare_model_data}.}

\item{parameters}{\code{list} of fixed and random model parameters.}
}
\value{
Object of class "naomi_fit".
}
\description{
Return minimised fit object. Often useful when saving the fit
object for later aggregation.
}
\keyword{internal}
