% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_survey_data.R
\name{find_circ_type}
\alias{find_circ_type}
\alias{reassign_survey_level}
\title{Define circumcision type}
\usage{
find_circ_type(survey_circumcision)

reassign_survey_level(survey_circumcision, areas, area_lev)
}
\arguments{
\item{survey_circumcision}{\itemize{
\item Information on male circumcision status from
surveys. If this is a list or contains more than one country, the function
is performed for each country present, returning a list.
}}

\item{areas}{\code{sf} shapefiles for specific country/region.}

\item{area_lev}{\itemize{
\item Desired admin boundary level to perform the analysis on.
}}
}
\value{
Surveys with circumcision type labelled by the column \code{type}.

Surveys with records reassigned to \code{area_lev}.
}
\description{
Using \code{circ_who} and \code{circ_where}, determines survey
type.

Join surveys with areas for each level leading up to the
desired area level. This distributes granular surveys to desired area_level
using parent_ids.
}
\keyword{internal}
