% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{threemc_oos_ppc}
\alias{threemc_oos_ppc}
\title{Posterior Predictive Distribution and checks on OOS survey}
\usage{
threemc_oos_ppc(
  fit,
  out,
  populations,
  survey_estimate,
  removed_years,
  type = "coverage",
  area_lev = 1,
  age_groups = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39",
    "40-44", "45-49", "50-54", "54-59", "0+", "10+", "15+", "10-24", "15-24", "10-29",
    "15-29", "10-39", "15-39", "10-49", "15-49"),
  CI_range = 0.95,
  N = 1000,
  compare_stats = TRUE
)
}
\arguments{
\item{fit}{Fit object returned by \link[naomi]{sample_tmb}, which includes,
among other things, the optimised parameters and subsequent sample for our
TMB model.}

\item{out}{Unaggregated results of model fitting outputted by
\link[threemc]{compute_quantiles}.}

\item{populations}{Single age male population counts by space and time.}

\item{survey_estimate}{Circumcision estimates based on empirical observations
from surveys.}

\item{removed_years}{Years removed from dataset used to fit model
(i.e. "withheld" surveys from OOS model fitting).}

\item{type}{The desired circumcision estimate. Can be one of
"probability", "incidence" or "coverage".}

\item{area_lev}{Area level you wish to aggregate to when performing posterior
predictive comparisons with survey estimates.}

\item{age_groups}{Age groups to aggregate by, Default:
c("0-4",   "5-9",   "10-14", "15-19", "20-24", "25-29",
"30-34", "35-39", "40-44", "45-49", "50-54", "54-59",
"0+",    "10+",   "15+",   "15-24", "10-24", 15-29",
"10-29", "15-39", "10-39", "15-49", "10-49")}

\item{CI_range}{CI interval about which you want to compare empirical and
posterior predictive estimates for left out surveys.}

\item{N}{Number of samples to generate, Default: 1000}

\item{compare_stats}{Set to TRUE if you wish to compute comparative
statistics (specifically, ELPD and CRPS) to compare with alternative models,
Default: TRUE}
}
\value{
\code{data.frame} with samples aggregated by \code{aggr_cols} and
weighted by population.
}
\description{
Aggregate specified \code{numeric} columns by population-weighted
age groups (rather than single year ages), split by specified categories.
}
